# app/vmselect/promql: improve performance of parseCache on systems with many CPU cores