@f41gh7 2024-11-28
Hello, sorry for delay. WDYT of making a parsing benchmark with enabled and disable cache? 

 Maybe it's not worth it at all to use this cache and it's better to parse expression each time.