@kiriklo 2024-12-04
Hello, I included benchmarks for `parsePromQLWithCache()` and `parsePromQLWithoutCache()`. Even with simple queries, there is a significant performance difference between them, as can be seen from the benchmarks below . So cache is here for a reason.

```
goos: windows
goarch: amd64
pkg: github.com/VictoriaMetrics/VictoriaMetrics/app/vmselect/promql
cpu: 11th Gen Intel(R) Core(TM) i9-11900K @ 3.50GHz
BenchmarkParsePromQLWithCacheSimple-8                    4441305               280.2 ns/op             0 B/op          0 allocs/op
BenchmarkParsePromQLWithoutCacheSimple-8                   40227             28816 ns/op           13040 B/op        375 allocs/op
BenchmarkParsePromQLWithCacheSimpleParallel-8            7435629               168.9 ns/op             0 B/op          0 allocs/op
BenchmarkParsePromQLWithoutCacheSimpleParallel-8           98808             12271 ns/op           13040 B/op        375 allocs/op
BenchmarkParsePromQLWithCacheComplex-8                   2962010               414.5 ns/op             0 B/op          0 allocs/op
BenchmarkParsePromQLWithoutCacheComplex-8                   3864            320808 ns/op          139867 B/op       2401 allocs/op
BenchmarkParsePromQLWithCacheComplexParallel-8           9242356               128.0 ns/op             0 B/op          0 allocs/op
BenchmarkParsePromQLWithoutCacheComplexParallel-8           5791            195799 ns/op          139882 B/op       2401 allocs/op
```