@f41gh7 2024-12-05
Thanks for the benchmark! We discussed it offline and had an assumption, that parser works fast enough and this cache may not be needed. Looks like it's not true and this cache and your optimization makes a lot of sense.