@hagen1778 2025-01-07
Hey @f41gh7 @kiriklo! Thanks for contribution!
While indeed sharded cache helps to scale better on multi-core systems, it only becomes visible on high rate of read operations. And one vmselect usually doesn't serve more than hundreds of queries per second.
And even if it will start serving thousands of distinct requests per second, the cache with 128*78 limit on number of elements will still remain inefficient, as elements will be purged and re-written too frequently.