@kiriklo 2025-01-07
Hello @hagen1778 , I agree with this. It might be helpful when vmselect restarts. Because this cache is not stored on disk after the restart, it will be filled again from zero, so it will reduce the impact of the component restart. Furthermore, the number of elements in the cache can be changed from a constant to a flag, which allows users with high select operations to tune the cache themselves.